+++ examples/fake_git.ml
@@ -30,7 +30,13 @@ let commit =
   let+ _amend = flag [ "amend" ] ~desc:"Amend a commit"
   and+ _all = flag [ "a" ] ~desc:"Commit all changes"
   and+ _branch = named_opt [ "b"; "branch" ] branch_conv
-  and+ _message = named_opt [ "m"; "message" ] string ~desc:"The commit message"
+  and+ _message =
+    named_opt
+      [ "m"; "message" ]
+      string
+      ~desc:
+        "The commit message. This description is extra long to exercise text wrapping in \
+         help messages."
   and+ _files = pos_all file ~desc:"The files to commit" in
   ()
 ;;